Address review comment: move methods to BazelProfile
