+++ analyzer/javatests/com/engflow/bazel/invocation/analyzer/bazelprofile/BazelProfileTestSuite.java
@@ -19,7 +19,6 @@ import org.junit.runners.Suite;
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
-  BazelProfileConstantsTest.class,
   BazelProfileTest.class,
 })
 public class BazelProfileTestSuite {}
